feat(ds-host): implement multiple auths per appspace user, including tsnetid
- change AppspaceUsersModel Create and Update functions
- implement add/remove of auths for appspace users
- disable auto-add shared users in ManageUsers appspace op
- add FullID to TSNetPeerUser
- implement UI to add/rm appspace user auths (needs more work)

+++ frontend-ds-host/src/views/ManageAppspace.vue
@@ -1,6 +1,7 @@
 <script lang="ts" setup>
 import { ref, Ref, reactive, computed, onMounted, onUnmounted, watch } from 'vue';
 
+import type { AppspaceUser } from '@/stores/types';
 import { useAppspacesStore } from '@/stores/appspaces';
 import { useAppspaceUsersStore } from '@/stores/appspace_users';
 import { useAppsStore } from '@/stores/apps';
@@ -137,7 +138,9 @@ async function tsnetSetConnect(connect:boolean) {
 }
 
 async function tsnetDeleteConfig() {
-	await appspacesStore.deleteTSNetData(props.appspace_id);
+	if( confirm("Delete tailscale node configuration data?") ) {
+		await appspacesStore.deleteTSNetData(props.appspace_id); 
+	}
 }
 
 const tsnet_peer_users = computed( () => {
@@ -145,6 +148,17 @@ const tsnet_peer_users = computed( () => {
 	if( pu === undefined ) return undefined;
 	return pu.value;
 });
+
+const tsnet_peer_matched_users = computed( () => {
+	const ret : Map<string,AppspaceUser> = new Map();
+	tsnet_peer_users.value?.forEach( (pu) => {
+		const u = appspaceUsersStore.findByAuth(props.appspace_id, 'tsnetid', pu.full_id);
+		if( u ) ret.set(pu.id, u);
+	});
+	return ret;
+});
+
+const show_tsnet_users = ref(false);
 
 </script>
 <template>
@@ -260,8 +274,11 @@ const tsnet_peer_users = computed( () => {
 				<div v-else-if="appspace.tsnet_data" class="px-4 sm:px-6 my-5">
 					<div v-if="appspace.tsnet_status.state == '' || appspace.tsnet_status.state == 'Off'">
 						<div class="flex justify-between">
-							<button @click.stop.prevent="tsnetDeleteConfig()" class="btn btn-blue">
-								Delete Configuration
+							<button @click.stop.prevent="tsnetDeleteConfig()" class="btn text-red-700">
+								<svg xmlns="http://www.w3.org/2000/svg" class="inline align-bottom h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
+									<path fill-rule="evenodd" d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z" clip-rule="evenodd" />
+								</svg>
+								<span class="hidden sm:inline-block">delete configuration</span>
 							</button>
 							<button @click.stop.prevent="tsnetSetConnect(true)" class="btn btn-blue">
 								Connect
@@ -278,6 +295,36 @@ const tsnet_peer_users = computed( () => {
 							<span v-if="appspace.tsnet_data.control_url ==''">(Tailscale)</span>
 							<span v-else>({{ appspace.tsnet_data.control_url }})</span>
 						</DataDef>
+						<DataDef field="Users:">
+							{{ tsnet_peer_matched_users.size }} of {{ tsnet_peer_users?.length || 0 }}
+							peers have access to this appspace
+							<a href="#" @click.stop.prevent="show_tsnet_users = !show_tsnet_users" class=btn>
+								{{ show_tsnet_users?"hide" : "show" }} peers
+							</a>
+						</DataDef>
+						<ul v-if="show_tsnet_users" class="border-gray-200 border-t my-4">
+							<li v-for="u in tsnet_peer_users" class="border-gray-200 border-b py-2 flex flex-col md:flex-row justify-between">
+								<span>
+									<p>
+										<span class="font-bold">{{ u.display_name }}</span>
+										({{ u.login_name }})
+									</p>
+									<p class="italic text-gray-500" v-if="u.sharee">appspace node was shared with them</p>
+									<p v-if="tsnet_peer_matched_users.has(u.id)">
+										Appspace user "{{ tsnet_peer_matched_users.get(u.id)?.display_name }}"
+										<router-link class="btn" :to="{name:'appspace-user', params:{appspace_id: props.appspace_id, proxy_id:tsnet_peer_matched_users.get(u.id)?.proxy_id}}">Edit</router-link>
+									</p>
+								</span>
+								<span v-if="tsnet_peer_matched_users.has(u.id)" class="justify-self-end">
+									matched
+									<!-- maybe link to manage user? -->
+									 <!-- maybe the user display name if different? -->
+								</span>
+								<span v-else class="justify-self-end">
+									add/match
+								</span>
+							</li>
+						</ul>
 						<div class="flex justify-end">
 							<button v-if="appspace.tsnet_data" @click.stop.prevent="tsnetSetConnect(!appspace.tsnet_data.connect)" class="btn btn-blue">
 								{{ appspace.tsnet_data.connect ? 'Disconnect' : 'Connect'}}
